+++ src/main/java/site/walkies/walkie/domain/internal/controller/Internal.java
@@ -2,11 +2,14 @@ package site.walkies.walkie.domain.internal.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
+import site.walkies.walkie.domain.egg.service.EggService;
+import site.walkies.walkie.domain.egg.service.dto.request.CreateEggRequest;
+import site.walkies.walkie.domain.egg.service.dto.response.EggResponse;
 import site.walkies.walkie.domain.spot.service.SpotPhotoService;
 import site.walkies.walkie.domain.spot.service.SpotSyncService;
+import site.walkies.walkie.global.auth.dto.MemberPrincipal;
 import site.walkies.walkie.global.web.dto.response.SuccessResponse;
 
 @RestController
@@ -16,6 +19,7 @@ public class Internal {
 
     private final SpotSyncService spotSyncService;
     private final SpotPhotoService spotPhotoService;
+    private final EggService eggService;
 
     @Operation(
             summary = "spot 생성 api",
@@ -45,5 +49,15 @@ public class Internal {
     public SuccessResponse<?> getSpotPhoto2() {
         spotPhotoService.enrichPhotosFromDetailUrls();
         return SuccessResponse.ok();
+    }
+
+    @Operation(
+            summary = "알 생성 test",
+            description = "사용자가 위치 정보를 보내면 100걸음짜리 test 알을 생성합니다."
+    )
+    @PostMapping("/egg")
+    public SuccessResponse<?> createTestEgg(@AuthenticationPrincipal MemberPrincipal memberPrincipal, @RequestBody CreateEggRequest createEggRequest) {
+        EggResponse response = eggService.createTestEgg(memberPrincipal.getMemberId(),createEggRequest.getLatitude(), createEggRequest.getLongitude());
+        return SuccessResponse.ok(response);
     }
 }
